Clean the repo by removing the deprecated modules and merging the files with the same functionalities. (#422)
