+++ src/agentscope/agents/__init__.py
@@ -5,7 +5,6 @@ from .operator import Operator
 from .dialog_agent import DialogAgent
 from .dict_dialog_agent import DictDialogAgent
 from .user_agent import UserAgent
-from .text_to_image_agent import TextToImageAgent
 from .rpc_agent import RpcAgent
 from .react_agent import ReActAgent
 from .rag_agent import LlamaIndexAgent
@@ -16,7 +15,6 @@ __all__ = [
     "Operator",
     "DialogAgent",
     "DictDialogAgent",
-    "TextToImageAgent",
     "UserAgent",
     "ReActAgent",
     "DistConf",
