+++ src/agentscope/models/ollama_model.py
@@ -3,9 +3,9 @@
 from abc import ABC
 from typing import Sequence, Any, Optional, List, Union, Generator
 
-from agentscope.message import Msg
-from agentscope.models import ModelWrapperBase, ModelResponse
-from agentscope.utils.tools import _convert_to_str
+from ..message import Msg
+from ..models import ModelWrapperBase, ModelResponse
+from ..utils.common import _convert_to_str
 
 try:
     import ollama
