+++ src/agentscope/prompt/__init__.py
@@ -6,11 +6,9 @@ from ._prompt_generator_zh import ChineseSystemPromptGenerator
 from ._prompt_generator_en import EnglishSystemPromptGenerator
 from ._prompt_comparer import SystemPromptComparer
 from ._prompt_optimizer import SystemPromptOptimizer
-from ._prompt_engine import PromptEngine
 
 
 __all__ = [
-    "PromptEngine",
     "SystemPromptGeneratorBase",
     "ChineseSystemPromptGenerator",
     "EnglishSystemPromptGenerator",
