+++ src/agentscope/rpc/__init__.py
@@ -8,7 +8,7 @@ try:
     from .rpc_agent_pb2_grpc import RpcAgentStub
     from .rpc_agent_pb2_grpc import add_RpcAgentServicer_to_server
 except ImportError as import_error:
-    from agentscope.utils.tools import ImportErrorReporter
+    from agentscope.utils.common import ImportErrorReporter
 
     RpcMsg = ImportErrorReporter(import_error, "distribute")  # type: ignore[misc]
     RpcAgentServicer = ImportErrorReporter(import_error, "distribute")
