--- src/agentscope/studio/static/html-drag-components/agent-texttoimageagent.html
@@ -1,28 +1,0 @@
-<div>
-    <div class="title-box" data-class="TextToImageAgent">
-        <div class="title-box-left-items">
-            <svg class="title-box-svg" viewBox="0 0 1024 1024"
-                 xmlns="http://www.w3.org/2000/svg">
-                <path d="M21.344 448h85.344v234.656H21.344V448zM554.656 192h64V106.656h-213.344V192h64v42.656h-320V896h725.344V234.656h-320V192z m234.688 128v490.656H234.688V320h554.656zM917.344 448h85.344v234.656h-85.344V448z"></path>
-                <path d="M341.344 512H448v106.656h-106.656V512zM576 512h106.656v106.656H576V512z"></path>
-            </svg>
-            <span>TextToImageAgent</span>
-        </div>
-        <button class="button copy-button">Copy</button>
-        <span class="toggle-arrow">&#x25B2;</span>
-    </div>
-    <div class="box">
-        <div class="readme">
-            Agent for text to image generation
-            <div>Node ID: <span class="node-id">ID_PLACEHOLDER</span></div>
-        </div>
-
-        <label>Name</label>
-        <input type="text" df-args-name placeholder="Assistant"
-               data-required="true">
-        <br>
-
-        <label>Model config name</label>
-        <input type="text" df-args-model_config_name placeholder="dall-e-3">
-    </div>
-</div>
