+++ src/agentscope/studio/templates/workstation.html
@@ -147,11 +147,6 @@
                                 UserAgent
                             </li>
                             <li class="workstation-sidebar-dragitem unselectable-text"
-                                data-node="TextToImageAgent"
-                                draggable="true" ondragstart="drag(event)">
-                                TextToImageAgent
-                            </li>
-                            <li class="workstation-sidebar-dragitem unselectable-text"
                                 data-node="DictDialogAgent"
                                 draggable="true" ondragstart="drag(event)">
                                 DictDialogAgent
