+++ src/agentscope/web/workstation/workflow_node.py
@@ -9,7 +9,6 @@ from agentscope import msghub
 from agentscope.agents import (
     DialogAgent,
     UserAgent,
-    TextToImageAgent,
     DictDialogAgent,
     ReActAgent,
 )
@@ -214,36 +213,6 @@ class UserAgentNode(WorkflowNode):
         return {
             "imports": "from agentscope.agents import UserAgent",
             "inits": f"{self.var_name} = UserAgent("
-            f"{kwarg_converter(self.opt_kwargs)})",
-            "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
-        }
-
-
-class TextToImageAgentNode(WorkflowNode):
-    """
-    A node representing a TextToImageAgent within a workflow.
-    """
-
-    node_type = WorkflowNodeType.AGENT
-
-    def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
-    ) -> None:
-        super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
-        self.pipeline = TextToImageAgent(**self.opt_kwargs)
-
-    def __call__(self, x: dict = None) -> dict:
-        return self.pipeline(x)
-
-    def compile(self) -> dict:
-        return {
-            "imports": "from agentscope.agents import TextToImageAgent",
-            "inits": f"{self.var_name} = TextToImageAgent("
             f"{kwarg_converter(self.opt_kwargs)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
             f"({DEFAULT_FLOW_VAR})",
@@ -840,7 +809,6 @@ NODE_NAME_MAPPING = {
     "Message": MsgNode,
     "DialogAgent": DialogAgentNode,
     "UserAgent": UserAgentNode,
-    "TextToImageAgent": TextToImageAgentNode,
     "DictDialogAgent": DictDialogAgentNode,
     "ReActAgent": ReActAgentNode,
     "Placeholder": PlaceHolderNode,
